+++ src/Network/Wai/SAML2/Assertion.hs
@@ -230,19 +230,24 @@ data AssertionAttribute = AssertionAttribute {
     attributeFriendlyName :: !(Maybe T.Text),
     -- | The name format.
     attributeNameFormat :: !T.Text,
-    -- | The value of the attribute.
-    attributeValue :: !T.Text
+    -- | The value of the attribute, concatened from the 'attributeValues'.
+    attributeValue :: !T.Text,
+    -- | The values of the attribute.
+    --
+    -- @since 0.7
+    attributeValues :: ![T.Text]
 } deriving (Eq, Show)
 
 instance FromXML AssertionAttribute where
     parseXML cursor = do
+        let attributeValues = cursor $/ element (saml2Name "AttributeValue") &/ content
         pure AssertionAttribute{
             attributeName = T.concat $ attribute "Name" cursor,
             attributeFriendlyName =
                 toMaybeText $ attribute "FriendlyName" cursor,
             attributeNameFormat = T.concat $ attribute "NameFormat" cursor,
-            attributeValue = T.concat $
-                cursor $/ element (saml2Name "AttributeValue") &/ content
+            attributeValue = T.concat attributeValues,
+            attributeValues = attributeValues
         }
 
 -- | SAML2 assertion statements (collections of assertion attributes).